feat: 공지사항 상세 메타데이터 설정
